+++ src/main/kotlin/emulator/architecture/instructions/set/ConvertToBase10.kt
@@ -14,7 +14,7 @@ class ConvertToBase10(
 ) : Instruction(nibbles) {
     lateinit var x: RX
     override fun processNibbles() {
-        val xValue = nibbles[0].toInt()
+        val xValue = nibbles[0].toUByte().toInt()
         x = r[xValue]    }
 
     override fun performOperation() {
