+++ src/main/kotlin/emulator/architecture/instructions/set/Jump.kt
@@ -9,9 +9,9 @@ class Jump(
 ) : Instruction(nibbles) {
     private lateinit var addressBytes: ByteArray
     override fun processNibbles() {
-        val highNibble = nibbles[0].toInt()
-        val middleNibble = nibbles[1].toInt()
-        val lowNibble = nibbles[2].toInt()
+        val highNibble = nibbles[0].toUByte().toInt()
+        val middleNibble = nibbles[1].toUByte().toInt()
+        val lowNibble = nibbles[2].toUByte().toInt()
 
         val address = (highNibble shl 8) or (middleNibble shl 4) or lowNibble
         addressBytes = Utili().intToByteArray(address)    }
