+++ src/main/kotlin/emulator/architecture/instructions/set/Read.kt
@@ -15,7 +15,7 @@ class Read(
     lateinit var x: RX
 
     override fun processNibbles() {
-        val xValue = nibbles[0].toInt()
+        val xValue = nibbles[0].toUByte().toInt()
         x = r[xValue]    }
 
     override fun performOperation() {
